NetworkBoundResource.kt

+++ app/src/main/java/com/example/android/weatherme/ui/current/CurrentViewModel.kt
@@ -3,6 +3,10 @@ package com.example.android.weatherme.ui.current
 import androidx.hilt.lifecycle.ViewModelInject
 import androidx.lifecycle.*
 import com.example.android.weatherme.data.Repository
+import com.example.android.weatherme.data.Resource
+import com.example.android.weatherme.data.Status
+import com.example.android.weatherme.data.database.entities.current.CurrentEntity
+import com.example.android.weatherme.data.network.api.Result
 import com.example.android.weatherme.utils.SingleLiveEvent
 import kotlinx.coroutines.launch
 
@@ -14,10 +18,14 @@ class CurrentViewModel @ViewModelInject constructor(
     val showSnackBarInt: SingleLiveEvent<Int> = SingleLiveEvent()
     private val loadNewCurrent: SingleLiveEvent<Long> = SingleLiveEvent()
 
-    val currentSelected = loadNewCurrent.switchMap { cityId ->
+    /*val currentSelected = loadNewCurrent.switchMap { cityId ->
         liveData {
             emitSource(repository.getCurrentByKey(cityId))
         }
+    }*/
+
+    val currentSelected: LiveData<Resource<CurrentEntity>> = loadNewCurrent.switchMap { cityId ->
+        repository.loadCurrent(cityId)
     }
 
     val perHour = loadNewCurrent.switchMap { cityId ->
@@ -27,17 +35,19 @@ class CurrentViewModel @ViewModelInject constructor(
     }
 
     val setShowData: LiveData<Boolean> = Transformations.map(currentSelected) { value ->
-        value?.let {
-            value.cityName?.isNotEmpty() ?: false
+        value.data?.let {
+            it.cityName?.isNotEmpty() ?: false
         }
     }
 
     fun deleteEntry() {
         viewModelScope.launch {
             currentSelected.value?.let {
-                val cityId = it.cityId
-                repository.deleteCurrent(cityId)
-                repository.deletePerHour(cityId)
+                if (it.status == Status.SUCCESS) {
+                    val cityId = it.data?.cityId ?: 0
+                    repository.deleteCurrent(cityId)
+                    repository.deletePerHour(cityId)
+                }
             }
         }
     }
